test: split configs
